Implement tree equality checking.
To test whether the tree structure is
correct, we can create a tree that doesn't
undergo any rotations, and then check
that the nodes all match up.

In the two trees, we compare them node by
node, ensuring that the two compared nodes
both have the same keys, the same parent,
and the same subtrees.