Remove the balance field from avl nodes
There are two ways to implement an AVL tree
in terms of what the nodes keep track of.
They can either keep track of their own
heights, or they can keep track of balances.

We may implement the balance version later,
but for now we focus on solely height tracking.