Use a void pointer to hide all info in header
With a void pointer, the client won't know even
the name of the implementation class, and we've
hidden the information as much as we possibly
can.

Whether this is good coding practice, I don't
actually know, because the calls are starting
to look somewhat ugly, with all the static casts.